requirements.txt
openai
langchain
langchain_groq
PyPDF2
python-dotenv
chromadb
python-telegram-bot
sentence-transformers
faiss-cpu
requests
flask
gunicorn
tqdm